collection summarize dir
