+++ aeon/classification/distance_based/_shape_dtw.py
@@ -8,6 +8,7 @@ import numpy as np
 
 # Tuning
 from sklearn.model_selection import GridSearchCV, KFold
+from sklearn.preprocessing import FunctionTransformer
 
 # Classifiers
 from aeon.classification.base import BaseClassifier
@@ -17,16 +18,11 @@ from aeon.classification.distance_based._time_series_neighbors import (
 from aeon.transformations.collection.dictionary_based._paa import PAA
 from aeon.transformations.collection.dwt import DWTTransformer
 
-# Done
+# Transformers
 from aeon.transformations.collection.hog1d import HOG1DTransformer
 from aeon.transformations.collection.segment import SlidingWindowSegmenter
 from aeon.transformations.collection.slope import SlopeTransformer
-from aeon.transformations.collection.summarize._extract import (
-    DerivativeSlopeTransformer,
-)
-
-# Transformers: To Do
-
+from aeon.utils.numba.general import slope_derivative_3d
 
 __author__ = ["vincent-nich12"]
 
@@ -359,7 +355,7 @@ class ShapeDTW(BaseClassifier):
                 return SlopeTransformer()
             return SlopeTransformer(num_intervals)
         elif tName == "derivative":
-            return DerivativeSlopeTransformer()
+            return FunctionTransformer(func=slope_derivative_3d)
         elif tName == "hog1d":
             return self._get_hog_transformer(parameters)
         else:
