+++ aeon/transformations/series/tests/test_plateaufinder.py
@@ -3,11 +3,11 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from aeon.transformations.collection.summarize import PlateauFinder
+from aeon.transformations.series.summarize import PlateauFinder
 
 
 @pytest.mark.parametrize("value", [np.nan, -10, 10, -0.5, 0.5])
-def test_PlateauFinder(value):
+def test_plateau_finder(value):
     # generate test data
     value = np.nan
     X = pd.DataFrame(
